Adapt rater
